Fixed SDF crash
Fixed the weird crash with SDFs, now works with both SDF and normal fonts.
Glyph layout info and the big texture atlas need to be re-added since I removed that for testing.
Also, remove the FontLoader as there is no benefit to justify the overhead of seperating loading from rendering and its dead code now.